Remove the pist fmt command
The fmt command formatted SQL files into pistachio's canonical form,
but its value over plain `dump` proved limited and the comment-aware
formatting work was abandoned. Drop the command and all of its
supporting code, tests, fixtures, scenario tests, and documentation.

formatSchemaSQL is moved to dump.go since it is now dump-only.
