+++ cmd/command/command_test.go
@@ -687,107 +687,3 @@ func TestApply_Run_ExecutedWithSkippedDrops(t *testing.T) {
 	assert.Less(t, addColPos, skippedPos, "executed DDL must precede skipped drop comment")
 	assert.NotContains(t, got, "-- No changes")
 }
-
-func TestFmt_Run(t *testing.T) {
-	tmpFile := filepath.Join(t.TempDir(), "test.sql")
-	require.NoError(t, os.WriteFile(tmpFile, []byte(`CREATE TABLE public.users (id integer NOT NULL, CONSTRAINT users_pkey PRIMARY KEY (id));`), 0o644))
-
-	client := pistachio.NewClient(&pistachio.Options{
-		Schemas: []string{"public"},
-	})
-
-	var buf bytes.Buffer
-	cmd := &command.Fmt{FmtOptions: pistachio.FmtOptions{Files: []string{tmpFile}}}
-	err := cmd.Run(client, &buf)
-	require.NoError(t, err)
-	assert.Contains(t, buf.String(), "CREATE TABLE public.users")
-	assert.Contains(t, buf.String(), "    id integer NOT NULL")
-}
-
-func TestFmt_Run_InvalidFile(t *testing.T) {
-	client := pistachio.NewClient(&pistachio.Options{
-		Schemas: []string{"public"},
-	})
-
-	var buf bytes.Buffer
-	cmd := &command.Fmt{FmtOptions: pistachio.FmtOptions{Files: []string{"/nonexistent/file.sql"}}}
-	err := cmd.Run(client, &buf)
-	require.Error(t, err)
-}
-
-func TestFmt_Run_WriteError(t *testing.T) {
-	tmpDir := t.TempDir()
-	readonlyDir := filepath.Join(tmpDir, "readonly")
-	require.NoError(t, os.MkdirAll(readonlyDir, 0o755))
-	inputFile := filepath.Join(readonlyDir, "test.sql")
-	require.NoError(t, os.WriteFile(inputFile, []byte(`CREATE TABLE public.users (id integer NOT NULL, CONSTRAINT users_pkey PRIMARY KEY (id));`), 0o444))
-	require.NoError(t, os.Chmod(readonlyDir, 0o555))
-	t.Cleanup(func() { os.Chmod(readonlyDir, 0o755) })
-
-	client := pistachio.NewClient(&pistachio.Options{
-		Schemas: []string{"public"},
-	})
-
-	var buf bytes.Buffer
-	cmd := &command.Fmt{FmtOptions: pistachio.FmtOptions{Files: []string{inputFile}, Write: true}}
-	err := cmd.Run(client, &buf)
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "failed to write")
-}
-
-func TestFmt_Run_Write(t *testing.T) {
-	tmpFile := filepath.Join(t.TempDir(), "test.sql")
-	require.NoError(t, os.WriteFile(tmpFile, []byte(`CREATE TABLE public.users (id integer NOT NULL, CONSTRAINT users_pkey PRIMARY KEY (id));`), 0o644))
-
-	client := pistachio.NewClient(&pistachio.Options{
-		Schemas: []string{"public"},
-	})
-
-	var buf bytes.Buffer
-	cmd := &command.Fmt{FmtOptions: pistachio.FmtOptions{Files: []string{tmpFile}, Write: true}}
-	err := cmd.Run(client, &buf)
-	require.NoError(t, err)
-	assert.Empty(t, buf.String())
-
-	content, err := os.ReadFile(tmpFile)
-	require.NoError(t, err)
-	assert.Contains(t, string(content), "CREATE TABLE public.users")
-	assert.Contains(t, string(content), "    id integer NOT NULL")
-}
-
-func TestFmt_Run_Check_Formatted(t *testing.T) {
-	// Already formatted file
-	tmpFile := filepath.Join(t.TempDir(), "test.sql")
-	formatted := "-- public.users\nCREATE TABLE public.users (\n    id integer NOT NULL,\n    CONSTRAINT users_pkey PRIMARY KEY (id)\n);\n"
-	require.NoError(t, os.WriteFile(tmpFile, []byte(formatted), 0o644))
-
-	client := pistachio.NewClient(&pistachio.Options{
-		Schemas: []string{"public"},
-	})
-
-	var buf bytes.Buffer
-	cmd := &command.Fmt{FmtOptions: pistachio.FmtOptions{Files: []string{tmpFile}, Check: true}}
-	err := cmd.Run(client, &buf)
-	require.NoError(t, err)
-	assert.Empty(t, buf.String())
-}
-
-func TestFmt_Run_Check_NotFormatted(t *testing.T) {
-	tmpFile := filepath.Join(t.TempDir(), "test.sql")
-	require.NoError(t, os.WriteFile(tmpFile, []byte(`CREATE TABLE public.users (id integer NOT NULL, CONSTRAINT users_pkey PRIMARY KEY (id));`), 0o644))
-
-	client := pistachio.NewClient(&pistachio.Options{
-		Schemas: []string{"public"},
-	})
-
-	var buf bytes.Buffer
-	cmd := &command.Fmt{FmtOptions: pistachio.FmtOptions{Files: []string{tmpFile}, Check: true}}
-	err := cmd.Run(client, &buf)
-	require.Error(t, err)
-
-	var notFormatted *command.ErrNotFormatted
-	require.ErrorAs(t, err, &notFormatted)
-	assert.Len(t, notFormatted.Files, 1)
-	assert.Contains(t, notFormatted.Error(), tmpFile)
-	assert.Contains(t, buf.String(), tmpFile)
-}
