--- test/scenario/testdata/fmt/unformatted2.sql
@@ -1,1 +1,0 @@
-CREATE TABLE public.products (id integer NOT NULL, name text NOT NULL, CONSTRAINT products_pkey PRIMARY KEY (id));
