--- testdata/fmt/domain.yml
@@ -1,6 +1,0 @@
-input: |
-  CREATE DOMAIN public.pos_int AS integer CONSTRAINT pos_check CHECK (VALUE > 0);
-expected: |
-  -- public.pos_int
-  CREATE DOMAIN public.pos_int AS integer
-      CONSTRAINT pos_check CHECK (value > 0);
