--- testdata/fmt/empty.yml
@@ -1,2 +1,0 @@
-input: ""
-expected: ""
