--- testdata/fmt/enum.yml
@@ -1,9 +1,0 @@
-input: |
-  CREATE TYPE public.status AS ENUM ('active','inactive','pending');
-expected: |
-  -- public.status
-  CREATE TYPE public.status AS ENUM (
-      'active',
-      'inactive',
-      'pending'
-  );
