--- testdata/fmt/execute.yml
@@ -1,13 +1,0 @@
-input: |
-  CREATE TABLE public.users (id integer NOT NULL, CONSTRAINT users_pkey PRIMARY KEY (id));
-  -- pist:execute SELECT true
-  CREATE OR REPLACE FUNCTION public.my_func() RETURNS void AS $$ BEGIN END; $$ LANGUAGE plpgsql;
-expected: |
-  -- public.users
-  CREATE TABLE public.users (
-      id integer NOT NULL,
-      CONSTRAINT users_pkey PRIMARY KEY (id)
-  );
-
-  -- pist:execute SELECT true
-  CREATE OR REPLACE FUNCTION public.my_func() RETURNS void AS $$ BEGIN END; $$ LANGUAGE plpgsql;
