--- testdata/fmt/idempotent.yml
@@ -1,34 +1,0 @@
-input: |
-  -- public.status
-  CREATE TYPE public.status AS ENUM (
-      'active',
-      'inactive'
-  );
-
-  -- public.users
-  CREATE TABLE public.users (
-      id integer NOT NULL,
-      name text NOT NULL,
-      CONSTRAINT users_pkey PRIMARY KEY (id)
-  );
-
-  -- public.active_users
-  CREATE OR REPLACE VIEW public.active_users AS
-  SELECT id, name FROM public.users WHERE name IS NOT NULL;
-expected: |
-  -- public.status
-  CREATE TYPE public.status AS ENUM (
-      'active',
-      'inactive'
-  );
-
-  -- public.users
-  CREATE TABLE public.users (
-      id integer NOT NULL,
-      name text NOT NULL,
-      CONSTRAINT users_pkey PRIMARY KEY (id)
-  );
-
-  -- public.active_users
-  CREATE OR REPLACE VIEW public.active_users AS
-  SELECT id, name FROM public.users WHERE name IS NOT NULL;
