--- testdata/fmt/mixed.yml
@@ -1,21 +1,0 @@
-input: |
-  CREATE TABLE public.users (id integer NOT NULL, name text NOT NULL, CONSTRAINT users_pkey PRIMARY KEY (id));
-  CREATE TYPE public.status AS ENUM ('active','inactive');
-  CREATE VIEW public.active_users AS SELECT id, name FROM public.users WHERE name IS NOT NULL;
-expected: |
-  -- public.status
-  CREATE TYPE public.status AS ENUM (
-      'active',
-      'inactive'
-  );
-
-  -- public.users
-  CREATE TABLE public.users (
-      id integer NOT NULL,
-      name text NOT NULL,
-      CONSTRAINT users_pkey PRIMARY KEY (id)
-  );
-
-  -- public.active_users
-  CREATE OR REPLACE VIEW public.active_users AS
-  SELECT id, name FROM public.users WHERE name IS NOT NULL;
