--- testdata/fmt/multiple_enums.yml
@@ -1,16 +1,0 @@
-input: |
-  CREATE TYPE public.status AS ENUM ('active', 'inactive');
-  CREATE TYPE public.role AS ENUM ('admin', 'user', 'guest');
-expected: |
-  -- public.status
-  CREATE TYPE public.status AS ENUM (
-      'active',
-      'inactive'
-  );
-
-  -- public.role
-  CREATE TYPE public.role AS ENUM (
-      'admin',
-      'user',
-      'guest'
-  );
