--- testdata/fmt/simple_table.yml
@@ -1,9 +1,0 @@
-input: |
-  CREATE TABLE public.users (id integer NOT NULL, name text NOT NULL, CONSTRAINT users_pkey PRIMARY KEY (id));
-expected: |
-  -- public.users
-  CREATE TABLE public.users (
-      id integer NOT NULL,
-      name text NOT NULL,
-      CONSTRAINT users_pkey PRIMARY KEY (id)
-  );
