--- testdata/fmt/view.yml
@@ -1,13 +1,0 @@
-input: |
-  CREATE TABLE public.users (id integer NOT NULL, CONSTRAINT users_pkey PRIMARY KEY (id));
-  CREATE VIEW public.active_users AS SELECT id FROM public.users;
-expected: |
-  -- public.users
-  CREATE TABLE public.users (
-      id integer NOT NULL,
-      CONSTRAINT users_pkey PRIMARY KEY (id)
-  );
-
-  -- public.active_users
-  CREATE OR REPLACE VIEW public.active_users AS
-  SELECT id FROM public.users;
